Refactored db_gateway and delete_previous.

+++ src/downloader/db_gateway.py
@@ -17,6 +17,7 @@
 # https://github.com/MiSTer-devel/Downloader_MiSTer
 
 from pathlib import Path
+from itertools import chain
 
 
 class DbGateway:
@@ -27,11 +28,22 @@ class DbGateway:
 
     def fetch_all(self, descriptions):
 
-        file_downloader = self._file_downloader_factory.create(parallel_update=True, silent=True, hash_check=False)
+        descriptions_by_file, local_files, remote_files = self._categorize_files_on_db_url(descriptions)
 
-        dbs = []
-        files = {}
-        errors = []
+        downloaded_files, download_errors = self._download_files(remote_files)
+
+        files_by_section = {descriptions_by_file[file]['section']: file for file in chain(downloaded_files, local_files)}
+
+        dbs, db_errors = self._read_dbs(descriptions, files_by_section)
+
+        self._cleanup_temp_files(descriptions_by_file)
+
+        return dbs, db_errors + self._identify_download_errors(download_errors, descriptions_by_file)
+
+    def _categorize_files_on_db_url(self, descriptions):
+        descriptions_by_file = {}
+        local_files = []
+        remote_files = []
 
         for section, description in descriptions.items():
             db_url = description['db_url']
@@ -40,33 +52,54 @@ class DbGateway:
                     db_url = self._file_system.resolve(db_url)
 
                 self._logger.debug('Loading db from local path: %s' % db_url)
-                dbs.append((section, self._file_system.load_db_from_file(db_url)))
+                descriptions_by_file[db_url] = description
+                local_files.append(db_url)
             else:
                 temp = self._file_system.temp_file()
-                files[temp] = description
+                descriptions_by_file[temp] = description
+                description[temp_marker] = True
                 self._logger.debug('Loading db from url: %s' % db_url)
-                file_downloader.queue_file({"url": db_url, "hash": "ignore", "size": 0}, temp)
+                remote_files.append((db_url, temp))
+
+        return descriptions_by_file, local_files, remote_files
+
+    def _download_files(self, remote_files):
+        file_downloader = self._file_downloader_factory.create(parallel_update=True, silent=True, hash_check=False)
+
+        for db_url, temp in remote_files:
+            file_downloader.queue_file({"url": db_url, "hash": "ignore", "size": 0}, temp)
 
         file_downloader.download_files(False)
 
-        correct_files = {files[file]['section']: file for file in file_downloader.correctly_downloaded_files()}
+        return file_downloader.correctly_downloaded_files(), file_downloader.errors()
 
+    def _read_dbs(self, descriptions, files_by_section):
+        dbs = []
+        errors = []
         for section, description in descriptions.items():
-            if section in correct_files:
+            if section in files_by_section:
                 try:
-                    db = self._file_system.load_db_from_file(correct_files[section], Path(description['db_url']).suffix.lower())
+                    db = self._file_system.load_db_from_file(files_by_section[section], Path(description['db_url']).suffix.lower())
                     dbs.append((section, db))
                 except Exception as e:
                     self._logger.debug(e)
                     self._logger.print('Could not load json from "%s"' % description['db_url'])
                     errors.append(description['db_url'])
 
-        for file in file_downloader.errors():
-            description = files[file]
-            self._logger.print('Could not download file from db_url: "%s"' % description['db_url'])
-            errors.append(description['db_url'])
-
-        for file in files:
-            self._file_system.unlink(file)
-
         return dbs, errors
+
+    def _cleanup_temp_files(self, descriptions_by_file):
+        for file, description in descriptions_by_file.items():
+            if temp_marker in description:
+                self._file_system.unlink(file)
+
+    def _identify_download_errors(self, download_errors, descriptions_by_file):
+        errors = [descriptions_by_file[file]['db_url'] for file in download_errors]
+
+        for db_url in errors:
+            self._logger.print('Could not download file from db_url: "%s"' % db_url)
+
+        return errors
+
+
+temp_marker = 'temp'
